+++ packages/core/features/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solana/wallet-standard-features
 
+## 1.1.0-alpha.5
+
+### Minor Changes
+
+-   Add `solana:signIn` (Sign In With Solana) feature
+
 ## 1.1.0-alpha.4
 
 ### Minor Changes